removed POST /api/groups
